requirements.txt
insightface
opencv-python
PySide6
onnxruntime-gpu
onnxruntime-openvino
dbus-python
numpy
psutil
requests
tqdm
pydantic
pydantic-core
onnx
cryptography
tpm2-pytss